refactor: simplify DI container and remove service interface abstraction
